% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmmixWire.R
\docType{data}
\name{expr.norm}
\alias{expr.norm}
\title{The goldenspike gene expression dataset}
\format{A data frame with 14010 rows (genes) and 6 variables (samples).}
\source{
http://www2.ccr.buffalo.edu/halfon/spike/spikedownloads.html
}
\usage{
data(expr.norm)
}
\description{
A dataset containing the goldenspike data "10a.dat".
}
\keyword{datasets}
