% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmmixWire.R
\name{pvalue.wire}
\alias{pvalue.wire}
\title{The p-values for the weighted contrast W_j.}
\usage{
pvalue.wire(wj, wj0)
}
\arguments{
\item{wj}{The weighted contrast W_j,  output of \code{\link{scores.wire}}.}

\item{wj0}{The null distribution for W_j,
output of \code{\link{wj2.permuted}}.}
}
\value{
A vector of p-values.
}
\description{
This function caculates the p-value of the weighted contrast,
W_j.
}
\details{
The p-values relate to the tests set by the
contrast in \code{\link{scores.wire}}.
}
\examples{


data(hedenlc)
dat<-hedenlc[1:100,] #for speed

set.seed(12345)
ret <-emmixwire(dat, g=3, ncov=3, nvcov=1, n1=7, n2=8, 
        debug=1, itmax=20, epsilon=1e-5)

###calculate the W_j
wj <- scores.wire(ret, contrast=c(1, -1))

### the null distribution of W_j
wj0 <- wj2.permuted(dat, ret, nB=19)
### the p-values of W_j
pv  <- pvalue.wire(wj, wj0)


}
\seealso{
\code{\link{emmixwire}} \code{\link{scores.wire}}
\code{\link{wj2.permuted}}.
}
\keyword{cluster}
\keyword{datasets}
