% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmmixWire.R
\name{scores.wire}
\alias{scores.wire}
\title{The weighted contrast}
\usage{
scores.wire(obj, contrast = NULL, useZ = TRUE)
}
\arguments{
\item{obj}{The return list of function emmixwire.}

\item{contrast}{The vector of the specified contrast.}

\item{useZ}{use the latent variable allocation Z_i (default,  TRUE)
or the posterior probability (FALSE).}
}
\value{
The vector of the statistic Wj
}
\description{
This function caculates the weighted contrast W_j 
in order to find out the most significant dfferentially expressed genes.
}
\details{
The number of classes of samples is either two or three.
}
\examples{

data(hedenlc)
hedenlc<-hedenlc[1:100,] #for speed
set.seed(123456)
obj<-emmixwire(hedenlc, g=5, ncov=3, nvcov=1, n1=7, n2=8, 
     debug=1, itmax=20, epsilon=1e-4) 
# calculate the weighted contrasts W_j
Wj  <- scores.wire(obj)

}
\seealso{
\code{\link{wire.init.fit}} \code{\link{scores.wire}}
}
\keyword{cluster}
\keyword{datasets}
